backend/requirements.txt
flask
python-dotenv
groq
transformers
torch
PyPDF2
flask-cors
sentence-transformers
